Use firebase with context defaults
