+++ src/App.test.tsx
@@ -1,5 +1,4 @@
-import {describe, it, expect} from "vitest";
-
+import { describe, it, expect } from "vitest";
 
 describe("App", () => {
   it("renders learn react link", () => {
