+++ src/App.tsx
@@ -7,7 +7,6 @@ import Home from "./components/upload/Home";
 import NotFound from "./components/NotFound";
 import Result from "./components/tree/Result";
 
-import FirebaseProvider from "./components/providers/FirebaseProvider";
 import FilesProvider from "./components/providers/FilesProvider";
 import { QueryClient, QueryClientProvider } from "react-query";
 
@@ -15,21 +14,19 @@ const queryClient = new QueryClient();
 
 const App: FC<{}> = () => {
   return (
-    <FirebaseProvider>
-      <QueryClientProvider client={queryClient}>
-        <FilesProvider>
-          <BrowserRouter>
-            <AppLayout>
-              <Routes>
-                <Route path="/" element={<Home />} />
-                <Route path="/tree/:treeId" element={<Result />} />
-                <Route path="*" element={<NotFound />} />
-              </Routes>
-            </AppLayout>
-          </BrowserRouter>
-        </FilesProvider>
-      </QueryClientProvider>
-    </FirebaseProvider>
+    <QueryClientProvider client={queryClient}>
+      <FilesProvider>
+        <BrowserRouter>
+          <AppLayout>
+            <Routes>
+              <Route path="/" element={<Home />} />
+              <Route path="/tree/:treeId" element={<Result />} />
+              <Route path="*" element={<NotFound />} />
+            </Routes>
+          </AppLayout>
+        </BrowserRouter>
+      </FilesProvider>
+    </QueryClientProvider>
   );
 };
