+++ src/hooks/useFirebase.ts
@@ -1,14 +1,8 @@
 import { useContext } from "react";
-import FirebaseContext from "../context/FirebaseContext";
+import { FirebaseContext } from "../context/FirebaseContext";
 
 const useFirebase = () => {
-  const firebase = useContext(FirebaseContext);
-  if (!firebase) {
-    throw new Error(
-      "useFirebase should be used within a FirebaseContext.Provider"
-    );
-  }
-  return firebase;
+  return useContext(FirebaseContext);
 };
 
 export default useFirebase;
